fix(curtailment): satisfy golangci-lint exhaustive / gosec / unconvert
CI's golangci-lint surfaced four issues on #185:

- exhaustive: EventState.IsTerminal switch was missing the non-terminal
  cases. Adding pending/active/restoring as an explicit no-match arm
  preserves the same return value while satisfying the linter.
- exhaustive: priorityName switch hid UNSPECIFIED / NORMAL / HIGH behind
  the default arm. Listing them explicitly documents that HIGH (rejected
  by the proto validator before this code runs) collapses to NORMAL.
- gosec G115: the uint32 -> int32 cast on candidate_min_power_w_override
  needed an explicit bounds check. The buf.validate ceiling already
  caps the value well below int32 max, but a defense-in-depth guard
  surfaces a clean InvalidArgument if interceptor wiring is ever
  bypassed instead of silently wrapping.
- unconvert: int32(*req.CandidateMinPowerWOverride) in the service was a
  no-op since the field is already *int32. Drop the conversion.

+++ server/internal/domain/curtailment/service.go
@@ -69,7 +69,7 @@ func (s *Service) Preview(ctx context.Context, req PreviewRequest) (*Plan, error
 	// service trusts that the override has cleared that gate.
 	minPowerW := orgConfig.CandidateMinPowerW
 	if req.CandidateMinPowerWOverride != nil {
-		minPowerW = int32(*req.CandidateMinPowerWOverride)
+		minPowerW = *req.CandidateMinPowerWOverride
 	}
 
 	// Cooldown bypass: EMERGENCY priority skips post_event_cooldown_sec.
